pyproject.toml
[tool.poetry]
name = "spip2md"
version = "0.0.1"
description = "Generate a static website with plain Markdown+YAML files from a SPIP CMS database"
license = "GPL-2.0"

authors = [
  "Guilhem Fauré <guilhem.faure@gfaure.eu>"
]

readme = ["README.md", "LICENSE"]
repository = "https://git.irsamc.ups-tlse.fr/gfaure/spip-db-2-md"
keywords = ["Markdown", "Static website", "SPIP", "Converter", "Exporter"]

classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
  "Operating System :: POSIX :: Linux",
  "Topic :: Text Processing"
]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
python-slugify = {extras = ["unidecode"], version = "^8.0.1"}
pymysql = "^1.0.3"
peewee = "^3.16.2"

[tool.poetry.scripts]
spip2md = "spip2md.lib:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
